mobilenet-infer-local/requirements.txt
onnxruntime
numpy
pillow
requests
